Refactor common code into common hook
